Check ins & reviews

Approved-by: Brooks G

+++ backend/src/main/java/com/techelevator/model/DestinationDao.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface DestinationDao {
 
 	
-	public List<Destination> getDestinationsByCategory(Integer category_id);
+	public List<Destination> getDestinationsByCategory(Long category_id);
 	
-	public Destination getDestination(Integer destination_id);
+	public Destination getDestination(Long destination_id);
 	
 	public List<Destination> getAllDestinations();
